+++ src/Ampersand/Output/ToJSON/Populations.hs
@@ -1,31 +1,15 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE MultiParamTypeClasses #-} 
 {-# LANGUAGE FlexibleInstances #-} 
-module Ampersand.Output.ToJSON.MySQLInstaller 
-  (MySQLInstaller,Populations)
+module Ampersand.Output.ToJSON.Populations 
+  (Populations)
 where
 import           Ampersand.ADL1
 import           Ampersand.Basics
-import           Ampersand.Output.ToJSON.JSONutils 
-import           Ampersand.Prototype.Generate
-import           Ampersand.Prototype.TableSpec
+import           Ampersand.Output.ToJSON.JSONutils
 import           Data.Maybe
 import qualified Data.Set as Set
 import qualified Data.Text as Text
-
-data MySQLInstaller = MySQLInstaller
-   { msiJSONallDBstructQueries :: [Text.Text]
-   , msiJSONallDefPopQueries   :: [Text.Text]
-   } deriving (Generic, Show)
-instance ToJSON MySQLInstaller where
-  toJSON = amp2Jason
-instance JSON MultiFSpecs MySQLInstaller where
- fromAmpersand _ multi = MySQLInstaller
-        { msiJSONallDBstructQueries = map queryAsSQL $ generateDBstructQueries fSpec False
-        , msiJSONallDefPopQueries = map queryAsSQL $ generateInitialPopQueries fSpec False
-        }
-  where
-    fSpec = userFSpec multi
 
 data Populations = Populations
    { epJSONatoms :: [AtomValuesOfConcept]