+++ static/zwolle/src/Ampersand/Rule/Conjunct.php
@@ -8,10 +8,12 @@
 namespace Ampersand\Rule;
 
 use Exception;
+use Generator;
 use Ampersand\Log\Logger;
 use Ampersand\Misc\Config;
 use Psr\Log\LoggerInterface;
 use Ampersand\Plugs\MysqlDB\MysqlDB;
+use Psr\Cache\CacheItemPoolInterface;
 
 /**
  *
@@ -27,6 +29,13 @@ class Conjunct
      * @var \Ampersand\Rule\Conjunct[]
      */
     private static $allConjuncts;
+
+    /**
+     * Cache pool that contains conjunct violations
+     *
+     * @var \Psr\Cache\CacheItemPoolInterface
+     */
+    protected static $conjunctCache;
     
     /**
      * Logger
@@ -41,7 +50,21 @@ class Conjunct
      * @var \Ampersand\Plugs\MysqlDB\MysqlDB
      */
     protected $database;
+
+    /**
+     * Undocumented variable
+     *
+     * @var \Psr\Cache\CacheItemPoolInterface
+     */
+    protected $cachePool;
     
+    /**
+     * Undocumented variable
+     *
+     * @var \Psr\Cache\CacheItemInterface
+     */
+    protected $cacheItem;
+
     /**
      * Conjunct identifier
      *
@@ -71,12 +94,11 @@ class Conjunct
     public $sigRuleNames;
     
     /**
-     * List of violation pairs
-     * [['src' => $srcAtom, 'tgt' => $tgtAtom]]
+     * Specifies if conjunct is already evaluated
      *
-     * @var array $conjunctViolations
+     * @var bool
      */
-    private $conjunctViolations = null;
+    protected $isEvaluated = false;
     
     /**
      * Conjunct constructor
@@ -85,7 +107,7 @@ class Conjunct
      * @param array $conjDef
      * @param \Psr\Log\LoggerInterface $logger
      */
-    private function __construct(array $conjDef, LoggerInterface $logger, MysqlDB $database)
+    private function __construct(array $conjDef, LoggerInterface $logger, MysqlDB $database, CacheItemPoolInterface $cachePool)
     {
         $this->logger = $logger;
 
@@ -95,6 +117,9 @@ class Conjunct
         $this->query = $conjDef['violationsSQL'];
         $this->invRuleNames = (array)$conjDef['invariantRuleNames'];
         $this->sigRuleNames = (array)$conjDef['signalRuleNames'];
+
+        $this->cachePool = $cachePool;
+        $this->cacheItem = $cachePool->getItem($this->id);
     }
     
     /**
@@ -158,64 +183,71 @@ class Conjunct
             return ($carry || in_array(substr($ruleName, 0, 3), ['UNI', 'INJ']));
         }, false);
     }
+
+    /**
+     * Get violation pairs of this conjunct
+     *
+     * @param boolean $forceReEvaluation
+     * @return array[] [['conjId' => '<conjId>', 'src' => '<srcAtomId>', 'tgt' => '<tgtAtomId>'], [], ..]
+     */
+    public function getViolations(bool $forceReEvaluation = false): array
+    {
+        // Skipping evaluation of UNI and INJ conjuncts. TODO: remove after fix for issue #535
+        if (Config::get('skipUniInjConjuncts', 'transactions') && $this->isUniOrInjConj()) {
+            $this->logger->debug("Skipping conjunct '{$this}', because it is part of a UNI/INJ rule");
+            return [];
+        }
+        
+        // If re-evaluation is forced
+        if ($forceReEvaluation || !$this->cacheItem->isHit()) {
+            $this->evaluate();
+            return $this->cacheItem->get();
+        }
+
+        // Otherwise get from cache
+        $this->logger->debug("Conjunct is already evaluated, getting violations from cache");
+        return $this->cacheItem->get();
+    }
     
     /**
      * Evaluate conjunct and return array with violation pairs
      *
-     * @param bool $fromCache
-     * @return array[] [['src' => '<srcAtomId>', 'tgt' => '<tgtAtomId>']]
+     * @return $this
      */
-    public function evaluate(bool $fromCache = true): array
+    public function evaluate(): Conjunct
     {
-        $this->logger->debug("Checking conjunct '{$this->id}' (fromCache:" . var_export($fromCache, true) . ")");
+        $this->logger->debug("Evaluating conjunct '{$this->id}'");
+        
         try {
-            // Skipping evaluation of UNI and INJ conjuncts. TODO: remove after fix for issue #535
-            if (Config::get('skipUniInjConjuncts', 'transactions') && $this->isUniOrInjConj()) {
-                $this->logger->debug("Skipping conjunct '{$this}', because it is part of a UNI/INJ rule");
-                return [];
-            } // If conjunct is already evaluated and conjunctCach may be used -> return violations
-            elseif (isset($this->conjunctViolations) && $fromCache) {
-                $this->logger->debug("Conjunct is already evaluated, getting violations from cache");
-                return $this->conjunctViolations;
-            } // Otherwise evaluate conjunct, cache and return violations
-            else {
-                // Execute conjunct query
-                $this->conjunctViolations = (array) $this->database->execute($this->getQuery());
-                
-                if (($count = count($this->conjunctViolations)) == 0) {
-                    $this->logger->debug("Conjunct '{$this->id}' holds");
-                } else {
-                    $this->logger->debug("Conjunct '{$this->id}' broken: {$count} violations");
-                }
+            // Execute conjunct query
+            $violations = array_map(function (array $pair) {
+                // Adds conjunct id to every pair
+                $pair['conjId'] = $this->id;
+                return $pair;
+            }, $this->database->execute($this->getQuery()));
 
-                return $this->conjunctViolations;
+            $this->isEvaluated = true;
+            $this->cacheItem->set($violations);
+            $this->cachePool->saveDeferred($this->cacheItem);
+            
+            if (($count = count($violations)) == 0) {
+                $this->logger->debug("Conjunct '{$this->id}' holds");
+            } else {
+                $this->logger->debug("Conjunct '{$this->id}' broken: {$count} violations");
             }
+
+            return $this;
         } catch (Exception $e) {
-            Logger::getUserLogger()->error("Error while checking conjunct '{$this->id}'");
+            Logger::getUserLogger()->error("Error while evaluating conjunct '{$this->id}'");
             $this->logger->error($e->getMessage());
-            return [];
+
+            return $this;
         }
     }
 
-    public function saveCache()
+    public function persistCacheItem()
     {
-        $dbsignalTableName = Config::get('dbsignalTableName', 'mysqlDatabase');
-
-        // Delete existing conjunct violation cache
-        $query = "DELETE FROM \"{$dbsignalTableName}\" WHERE \"conjId\" = '{$this->id}'";
-        $this->database->execute($query);
-        
-        // Save new violations (if any)
-        if (!empty($this->conjunctViolations)) {
-            // Add new conjunct violation to database
-            $query = "INSERT IGNORE INTO \"{$dbsignalTableName}\" (\"conjId\", \"src\", \"tgt\") VALUES ";
-            $values = [];
-            foreach ($this->conjunctViolations as $violation) {
-                $values[] = "('{$this->id}', '" . $this->database->escape($violation['src']) . "', '" . $this->database->escape($violation['tgt']) . "')";
-            }
-            $query .= implode(',', $values);
-            $this->database->execute($query);
-        }
+        $this->cachePool->save($this->cacheItem);
     }
     
     /**********************************************************************************************
@@ -224,6 +256,37 @@ class Conjunct
      *
      *********************************************************************************************/
     
+    /**
+     * Get conjunct violations (if possible from cache) for given set of conjuncts
+     *
+     * @param \Ampersand\Rule\Conjunct[] $conjuncts
+     * @return \Generator
+     */
+    public static function getConjunctViolations(array $conjuncts = []): Generator
+    {
+        // Foreach conjunct provided, check if there is a hit in cache (i.e. ->isHit())
+        $hits = $nonHits = [];
+        foreach ($conjuncts as $conjunct) {
+            /** @var \Ampersand\Rule\Conjunct $conjunct */
+            if ($conjunct->cacheItem->isHit()) {
+                $hits[] = $conjunct->id;
+            } else {
+                $nonHits[] = $conjunct;
+            }
+        }
+
+        // For all hits, use CachPoolInterface->getItems()
+        foreach (self::$conjunctCache->getItems($hits) as $cacheItem) {
+            /** @var \Psr\Cache\CacheItemInterface $cacheItem */
+            yield from $cacheItem->get();
+        }
+
+        // For all non-hits, get violations from Conjunct object
+        foreach ($nonHits as $conjunct) {
+            yield from $conjunct->getViolations();
+        }
+    }
+
     /**
      * Return conjunct object
      *
@@ -262,14 +325,15 @@ class Conjunct
      * @param \Ampersand\Plugs\MysqlDB\MysqlDB $database
      * @return void
      */
-    public static function setAllConjuncts(string $fileName, LoggerInterface $logger, MysqlDB $database)
+    public static function setAllConjuncts(string $fileName, LoggerInterface $logger, MysqlDB $database, CacheItemPoolInterface $cachePool)
     {
         self::$allConjuncts = [];
+        self::$conjunctCache = $cachePool;
         
         $allConjDefs = (array)json_decode(file_get_contents($fileName), true);
     
         foreach ($allConjDefs as $conjDef) {
-            self::$allConjuncts[$conjDef['id']] = new Conjunct($conjDef, $logger, $database);
+            self::$allConjuncts[$conjDef['id']] = new Conjunct($conjDef, $logger, $database, $cachePool);
         }
     }
 }